final version, new subject.

+++ module-01/ex1/ft_garden_data.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 class Plant():
-    def __init__(self, name: str, height: int, age: int) -> None:
+    def __init__(self, name: str, height: float, age: int) -> None:
         self.name = name
         self.height = height
         self.age = age
